[tlo-parsing] Move TLNode to header.
