+++ common/tlo-parsing/tl-dependency-graph.cpp
@@ -11,52 +11,6 @@
 #include "common/algorithms/find.h"
 #include "common/tlo-parsing/tl-objects.h"
 
-namespace vk {
-namespace tlo_parsing {
-
-struct TLNode {
-  enum {
-    holds_combinator,
-    holds_type,
-  } node_type;
-
-  explicit TLNode(const combinator *c) :
-    node_type(holds_combinator) {
-    tl_object_ptr_.combinator_ptr = c;
-  }
-
-  explicit TLNode(const type *t) :
-    node_type(holds_type) {
-    tl_object_ptr_.type_ptr = t;
-  }
-
-  bool is_type() const {
-    return node_type == holds_type;
-  }
-
-  bool is_combinator() const {
-    return node_type == holds_combinator;
-  }
-
-  const type *get_type() const {
-    return tl_object_ptr_.type_ptr;
-  }
-
-  const combinator *get_combinator() const {
-    return tl_object_ptr_.combinator_ptr;
-  }
-
-private:
-  // TODO when we switch to C++17, use std::variant
-  union {
-    const combinator *combinator_ptr;
-    const type *type_ptr;
-  } tl_object_ptr_;
-};
-
-} // namespace tlo_parsing
-} // namespace vk
-
 vk::tlo_parsing::DependencyGraph::DependencyGraph() = default;
 
 vk::tlo_parsing::DependencyGraph::DependencyGraph(vk::tlo_parsing::tl_scheme *scheme) : scheme(scheme) {
